Start moving new_filled_column
